Move cancel code to service, refactor redux code
