+++ src/applications/vaos/appointment-list/components/ConfirmedAppointmentDetailsPage/index.jsx
@@ -29,7 +29,7 @@ import ErrorMessage from '../../../components/ErrorMessage';
 import FullWidthLayout from '../../../components/FullWidthLayout';
 import Breadcrumbs from '../../../components/Breadcrumbs';
 import {
-  cancelAppointment,
+  startAppointmentCancel,
   closeCancelAppointment,
   confirmCancelAppointment,
   fetchConfirmedAppointmentDetails,
@@ -294,7 +294,9 @@ export default function ConfirmedAppointmentDetailsPage() {
                       className="fas fa-times vads-u-margin-right--1 vads-u-font-size--lg"
                     />
                     <button
-                      onClick={() => dispatch(cancelAppointment(appointment))}
+                      onClick={() =>
+                        dispatch(startAppointmentCancel(appointment))
+                      }
                       aria-label={`Cancel appointment on ${formatAppointmentDate(
                         moment.parseZone(appointment.start),
                       )}`}
